feat(BodyPhysics): body physics based on CharacterController
The Unity `CharacterController` was made for non-VR usage but is
equally helpful when implementing the concept of "body physics" for
VR. Three new components have been added that allow setting the
height of the controller to match a given Game Object, moving a Game
Object to be at the controller's top and most importantly moving the
controller by specifying a movement from one Game Object to the
other to replicate. With these components a body representation in
form of the controller is able to follow and match a VR headset
moving in the real world while adhering to the virtual world's
constraints imposed by both the physical world as well as other
components affecting the controller.