feat(Tracking): allow emitting TransformData from GameObject
`TransformData` is used throughout the codebase to offer a more
powerful representation of a `Transform`. The need to create an
instance of the data type from a given Game Object arises in
multiple places so a new component has been added to allow doing
that without having to resort to do so in the script that needs it.
The new component allows setting the source Game Object to create
the `TransformData` for as well as emitting the event with that data
manually.